backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
opencv-python-headless
easyocr
python-bidi==0.4.2
pytesseract
pyzbar
rapidfuzz
python-dotenv
anthropic
httpx
